+++ apps/seller/src/widgets/reset-button/index.ts
@@ -1,0 +1,1 @@
+export { ResetButton } from './reset-button'
